requirements.txt
python-dotenv
fastapi
uvicorn[standard]
teams-ai>=1.6.0,<2.0.0